Dummy filterbox added

+++ src/utils/FormateValues.js
@@ -5,8 +5,8 @@ export function formatCurrency(value) {
 export function formatDate(dateString) {
     const date = new Date(dateString);
     const day = String(date.getDate()).padStart(2, '0'); // Ensure two digits
-    const month = date.toLocaleString('en-US', { month: 'long' }); // Get full month name
+    const month = date.toLocaleString('en-US', { month: 'short' }); // Get full month name
     const year = date.getFullYear();
 
-    return `${day} ${month},${year}`;
+    return `${day} ${month}, ${year}`;
 }
